+++ libraries/TCA9555/TCA9555.cpp
@@ -1,7 +1,7 @@
 //
 //    FILE: TCA9555.cpp
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.4.0
+// VERSION: 0.4.1
 // PURPOSE: Arduino library for I2C TCA9555 16 channel port expander
 //    DATE: 2021-06-09
 //     URL: https://github.com/RobTillaart/TCA9555
@@ -10,7 +10,7 @@
 #include "TCA9555.h"
 
 
-//  REGISTERS
+//  REGISTERS P23-24
 #define TCA9555_INPUT_PORT_REGISTER_0     0x00    //  read()
 #define TCA9555_INPUT_PORT_REGISTER_1     0x01
 #define TCA9555_OUTPUT_PORT_REGISTER_0    0x02    //  write()
@@ -30,7 +30,7 @@ TCA9555::TCA9555(uint8_t address, TwoWire *wire)
 }
 
 
-bool TCA9555::begin(bool mode, uint16_t mask)
+bool TCA9555::begin(uint8_t mode, uint16_t mask)
 {
   if ((_address < 0x20) || (_address > 0x27)) return false;
   if (! isConnected()) return false;
@@ -366,9 +366,21 @@ uint8_t TCA9555::readRegister(uint8_t reg)
 
 /////////////////////////////////////////////////////////////////////////////
 //
-//  TCA9535
+//  DERIVED CLASSES TCA9535 PCA9555 PCA9535
 //
 TCA9535::TCA9535(uint8_t address, TwoWire *wire)
+        :TCA9555(address, wire)
+{
+  _type = 35;
+}
+
+PCA9555::PCA9555(uint8_t address, TwoWire *wire)
+        :TCA9555(address, wire)
+{
+  _type = 55;
+}
+
+PCA9535::PCA9535(uint8_t address, TwoWire *wire)
         :TCA9555(address, wire)
 {
   _type = 35;